+++ Sources/OnboardingKit/_Deprecated/TutorialSlideView.swift
@@ -1,0 +1,11 @@
+#if os(iOS)
+import Foundation
+
+@available(iOS 14.0, *)
+@available(*, deprecated, renamed: "TutorialPageView")
+public typealias TutorialSlideView = TutorialPageView
+
+@available(iOS 14.0, *)
+@available(*, deprecated, renamed: "TutorialPageViewStyle")
+public typealias TutorialSlideViewStyle = TutorialPageViewStyle
+#endif